+++ Editor/Processors/DeleteGameObjectProcessor.cs
@@ -1,3 +1,4 @@
+using Anatawa12.AvatarOptimizer.ErrorReporting;
 using UnityEngine;
 
 namespace Anatawa12.AvatarOptimizer.Processors
@@ -6,7 +7,7 @@ namespace Anatawa12.AvatarOptimizer.Processors
     {
         public void Process(OptimizerSession session)
         {
-            foreach (var mergePhysBone in session.GetComponents<DeleteGameObject>())
+            BuildReport.ReportingObjects(session.GetComponents<DeleteGameObject>(), mergePhysBone =>
             {
                 void Destroy(Object obj)
                 {
@@ -25,7 +26,7 @@ namespace Anatawa12.AvatarOptimizer.Processors
                         Destroy(component);
                     return true;
                 });
-            }
+            });
         }
     }
 }