+++ Editor/Processors/MakeChildrenProcessor.cs
@@ -1,15 +1,17 @@
+using Anatawa12.AvatarOptimizer.ErrorReporting;
+
 namespace Anatawa12.AvatarOptimizer.Processors
 {
     internal class MakeChildrenProcessor
     {
         public void Process(OptimizerSession session)
         {
-            foreach (var makeChildren in session.GetComponents<MakeChildren>())
+            BuildReport.ReportingObjects(session.GetComponents<MakeChildren>(), makeChildren =>
             {
                 foreach (var makeChildrenChild in makeChildren.children.GetAsSet())
                     if (makeChildrenChild)
                         makeChildrenChild.parent = makeChildren.transform;
-            }
+            });
         }
     }
 }